fix commenting
I got errors when using comment-nvim and I think this should fix it. I haven't tried it on neovim-basic-ide but this works in my personal config and it matches the documentation. There are no open issues related to the commenting, so I am not sure if this is a problem only for me